requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1
reportlab==4.1.0
google-api-python-client==2.121.0
google-cloud-bigquery==3.18.0
google-auth==2.28.1
python-dotenv==1.0.1
pytest==8.0.2
httpx==0.27.0
pytest-mock==3.12.0
qrcode[pil]
Pillow
